f64 -> f32

+++ src/lorenz/attractor.rs
@@ -1,12 +1,12 @@
-const A: f64 = 28.; 
-const B: f64 = 10.;
-const C: f64 = 2.67;
-const DT: f64 = 0.01;
+const A: f32 = 28.; 
+const B: f32 = 10.;
+const C: f32 = 2.67;
+const DT: f32 = 0.01;
 
-pub fn generate(x: &f64, y: &f64, z: &f64) -> (f64, f64, f64) {
-    let dx: f64 = (B * (y - x)) * DT;
-    let dy: f64 = (x * (A - z) - y) * DT;
-    let dz: f64 = ((x * y) - (C * z)) * DT;
+pub fn generate(x: &f32, y: &f32, z: &f32) -> (f32, f32, f32) {
+    let dx: f32 = (B * (y - x)) * DT;
+    let dy: f32 = (x * (A - z) - y) * DT;
+    let dz: f32 = ((x * y) - (C * z)) * DT;
 
     return (dx, dy, dz);
 }
